+++ packages/dev/src/index.ts
@@ -41,7 +41,7 @@
  * ## API
  *
  * - Vite: {@link viteFaasJsServer}
- * - Test: {@link test}, {@link FuncWarper}, {@link streamToString}
+ * - Test: {@link test}, {@link FuncWarper}, {@link streamToText}, {@link streamToObject}, {@link streamToString}
  *
  * @packageDocumentation
  */