+++ packages/dev/src/typegen.ts
@@ -1,8 +1,8 @@
 import { existsSync } from 'node:fs'
 import { mkdir, readdir, readFile, writeFile } from 'node:fs/promises'
 import { dirname, join, relative } from 'node:path'
-import { loadConfig } from '@faasjs/load'
 import { Logger } from '@faasjs/logger'
+import { loadConfig } from '@faasjs/node-utils'
 import { resolveServerConfig } from './server_config'
 
 type RouteTypeItem = {