+++ packages/load/src/__tests__/index.test.ts
@@ -1,6 +1,0 @@
-import { expect, it } from 'vitest'
-import * as load from '../index'
-
-it('should work', () => {
-  expect(load).toHaveProperty('loadConfig')
-})