+++ packages/node-utils/src/__tests__/deep_merge.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { deepMerge } from '..'
+import { deepMerge } from '../deep_merge'
 
 describe('deepMerge', () => {
   it('should work', () => {