+++ packages/node-utils/src/__tests__/index.test.ts
@@ -1,0 +1,13 @@
+import { expect, it } from 'vitest'
+import * as nodeUtils from '../index'
+
+it('should export helpers', () => {
+  expect(nodeUtils).toHaveProperty('deepMerge')
+  expect(nodeUtils).toHaveProperty('loadConfig')
+  expect(nodeUtils).toHaveProperty('loadFunc')
+  expect(nodeUtils).toHaveProperty('loadPackage')
+  expect(nodeUtils).toHaveProperty('detectNodeRuntime')
+  expect(nodeUtils).toHaveProperty('streamToText')
+  expect(nodeUtils).toHaveProperty('streamToObject')
+  expect(nodeUtils).toHaveProperty('streamToString')
+})