+++ packages/server/src/server/__tests__/funcs/runtime.func.ts
@@ -1,6 +1,6 @@
 import { Func } from '@faasjs/func'
 import { Http } from '@faasjs/http'
-import { detectNodeRuntime } from '@faasjs/load'
+import { detectNodeRuntime } from '@faasjs/node-utils'
 
 const http = new Http()